requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
youtube-transcript-api==0.6.2
google-generativeai==0.7.2
openai>=1.40
webvtt-py==0.4.6
yt-dlp==2024.08.06
streamlit==1.37.1
pandas==2.2.2
numpy==2.0.1
